+++ R/utils.R
@@ -25,3 +25,12 @@ knit_devel <- function(input, ...) {
                     output_options = list(devel = TRUE),
                     quiet = TRUE)
 }
+
+
+is_available <- function(package, version = NULL) {
+  installed <- nzchar(system.file(package = package))
+  if (is.null(version)) {
+    return(installed)
+  }
+  installed && isTRUE(utils::packageVersion(package) >= version)
+}